requirements.txt
taipy-gui
openai
python-dotenv